Move model to cases app, rename model and remove wrapper
